# server/gen/swagger/swagger.yaml
basePath: /api/v1
definitions:
  apierror.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  document.CreateDocumentRequest:
    properties:
      branch_name:
        maxLength: 100
        type: string
      content:
        items:
          type: integer
        type: array
      document_type:
        type: string
      folder_id:
        type: string
      is_template:
        type: boolean
      language:
        type: string
      slug:
        maxLength: 255
        type: string
      story_id:
        type: string
      summary:
        type: string
      title:
        maxLength: 500
        type: string
    required:
    - branch_name
    - language
    - slug
    - story_id
    - title
    type: object
  document.DocumentResponse:
    properties:
      branch_name:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      document_type:
        type: string
      folder_id:
        type: string
      id:
        type: string
      is_locked:
        type: boolean
      is_template:
        type: boolean
      language:
        type: string
      previous_version_id:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/document.DocumentStatus'
      story_id:
        type: string
      summary:
        type: string
      title:
        type: string
      total_downloads:
        type: integer
      total_shares:
        type: integer
      total_views:
        type: integer
      total_words:
        type: integer
      updated_at:
        type: string
      version_number:
        type: integer
    type: object
  document.DocumentStatus:
    enum:
    - draft
    - pending_review
    - is_reviewed
    - is_approved
    - rejected
    - archived
    - deleted
    type: string
    x-enum-varnames:
    - DocumentStatusDraft
    - DocumentStatusPendingReview
    - DocumentStatusIsReviewed
    - DocumentStatusIsApproved
    - DocumentStatusRejected
    - DocumentStatusArchived
    - DocumentStatusDeleted
  document.UpdateDocumentRequest:
    properties:
      branch_name:
        maxLength: 100
        type: string
      content:
        items:
          type: integer
        type: array
      document_type:
        type: string
      folder_id:
        type: string
      is_locked:
        type: boolean
      is_template:
        type: boolean
      language:
        type: string
      slug:
        maxLength: 255
        type: string
      summary:
        type: string
      title:
        maxLength: 500
        type: string
    type: object
...
  sqlc.DocumentStatus:
    enum:
    - draft
    - pending_review
    - is_reviewed
    - is_approved
    - rejected
    - archived
    - deleted
    type: string
    x-enum-varnames:
    - DocumentStatusDraft
    - DocumentStatusPendingReview
    - DocumentStatusIsReviewed
    - DocumentStatusIsApproved
    - DocumentStatusRejected
    - DocumentStatusArchived
    - DocumentStatusDeleted
  sqlc.DocumentStatusHistory:
    properties:
      content:
        type: string
      created_at:
        type: string
      document_id:
        type: string
      id:
        type: string
      set_status_by:
        type: string
      status:
        $ref: '#/definitions/sqlc.DocumentStatus'
      updated_at:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: trao0312@gmail.com
    name: API Support
    url: https://github.com/justblue/samsa
  description: Samsa is a simple backend for a writing platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Samsa
  version: 0.1.0
paths:
...
  /documents:
    get:
      consumes:
      - application/json
      description: Retrieves documents based on filters.
      parameters:
      - description: Story UUID
        in: query
        name: story_id
        type: string
      - description: Folder UUID
        in: query
        name: folder_id
        type: string
      - description: Document status
        in: query
        name: status
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/document.DocumentResponse'
            type: array
      summary: List documents
      tags:
      - documents
    post:
      consumes:
      - application/json
      description: Creates a new document. Requires `user actor`.
      parameters:
      - description: Document creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/document.CreateDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/document.DocumentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.APIError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      security:
      - BearerAuth: []
      summary: Create document
      tags:
      - documents
  /documents/{document_id}:
    delete:
      description: Soft deletes a document. Requires `user actor` and ownership.
      parameters:
      - description: Document UUID
        in: path
        name: document_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete document
      tags:
      - documents
    get:
      consumes:
      - application/json
      description: Retrieves a document by its UUID.
      parameters:
      - description: Document UUID
        in: path
        name: document_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/document.DocumentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
      security:
      - BearerAuth: []
      summary: Get document by ID
      tags:
      - documents
    patch:
      consumes:
      - application/json
      description: Updates document details. Requires `user actor` and ownership.
      parameters:
      - description: Document UUID
        in: path
        name: document_id
        required: true
        type: string
      - description: Document update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/document.UpdateDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/document.DocumentResponse'
      security:
      - BearerAuth: []
      summary: Update document
      tags:
      - documents
  /documents/{document_id}/approve:
    post:
      consumes:
      - application/json
      description: Approves a document. Requires approval permissions.
      parameters:
      - description: Document UUID
        in: path
        name: document_id
        required: true
        type: string
      - description: Comments
        in: body
        name: request
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/document.DocumentResponse'
      security:
      - BearerAuth: []
      summary: Approve document
      tags:
      - documents
  /documents/{document_id}/archive:
    post:
      description: Archives a document. Requires `user actor` and ownership.
      parameters:
      - description: Document UUID
        in: path
        name: document_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/document.DocumentResponse'
      security:
      - BearerAuth: []
      summary: Archive document
      tags:
      - documents
  /documents/{document_id}/reject:
    post:
      consumes:
      - application/json
      description: Rejects a document. Requires approval permissions.
      parameters:
      - description: Document UUID
        in: path
        name: document_id
        required: true
        type: string
      - description: Comments
        in: body
        name: request
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/document.DocumentResponse'
      security:
      - BearerAuth: []
      summary: Reject document
      tags:
      - documents
  /documents/{document_id}/status-history:
    get:
      description: Retrieves status history for a document.
      parameters:
      - description: Document UUID
        in: path
        name: document_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sqlc.DocumentStatusHistory'
            type: array
      summary: Get document status history
      tags:
      - documents
  /documents/{document_id}/submit:
    post:
      description: Submits a document for review. Requires `user actor` and ownership.
      parameters:
      - description: Document UUID
        in: path
        name: document_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/document.DocumentResponse'
      security:
      - BearerAuth: []
      summary: Submit document for review
      tags:
      - documents
  /documents/{document_id}/versions:
    get:
      description: Retrieves version history for a document.
      parameters:
      - description: Document UUID
        in: path
        name: document_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/document.DocumentResponse'
            type: array
      summary: Get document version history
      tags:
      - documents
  /documents/{document_id}/view:
    post:
      description: Increments the view count for a document.
      parameters:
      - description: Document UUID
        in: path
        name: document_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/document.DocumentResponse'
      summary: Increment document views
      tags:
      - documents
  /documents/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Retrieves a document by its slug.
      parameters:
      - description: Document slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/document.DocumentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Get document by slug
      tags:
      - documents
...
swagger: "2.0"